+++ packages/schema/test/TreeFormatter.test.ts
@@ -1,7 +1,7 @@
 import * as ParseResult from "@effect/schema/ParseResult"
 import * as S from "@effect/schema/Schema"
 import * as Util from "@effect/schema/test/util"
-import * as _ from "@effect/schema/TreeFormatter"
+import * as TreeFormatter from "@effect/schema/TreeFormatter"
 import * as Context from "effect/Context"
 import * as Effect from "effect/Effect"
 import * as Either from "effect/Either"
@@ -637,7 +637,7 @@ describe("TreeFormatter", () => {
     const result = S.decodeUnknownEither(Name)("")
 
     // no service
-    expect(Either.mapLeft(result, (error) => Effect.runSync(_.formatError(error))))
+    expect(Either.mapLeft(result, (error) => Effect.runSync(TreeFormatter.formatError(error))))
       .toStrictEqual(Either.left("Invalid string"))
 
     // it locale
@@ -646,7 +646,7 @@ describe("TreeFormatter", () => {
         result,
         (error) =>
           Effect.runSync(
-            _.formatError(error).pipe(Effect.provideService(Translator, {
+            TreeFormatter.formatError(error).pipe(Effect.provideService(Translator, {
               locale: "it",
               translations
             }))
@@ -660,7 +660,7 @@ describe("TreeFormatter", () => {
         result,
         (error) =>
           Effect.runSync(
-            _.formatError(error).pipe(Effect.provideService(Translator, {
+            TreeFormatter.formatError(error).pipe(Effect.provideService(Translator, {
               locale: "en",
               translations
             }))