+++ packages/platform-node/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @effect/platform-node
 
+## 0.45.28
+
+### Patch Changes
+
+- Updated dependencies [[`365a486`](https://github.com/Effect-TS/effect/commit/365a4865de5e47ce09f4cfd51fc0f67438f82a57)]:
+  - @effect/platform@0.48.26
+  - @effect/platform-node-shared@0.3.26
+
+## 0.45.27
+
+### Patch Changes
+
+- Updated dependencies [[`dadc690`](https://github.com/Effect-TS/effect/commit/dadc6906121c512bc32be22b52adbd1ada834594), [`d209171`](https://github.com/Effect-TS/effect/commit/d2091714a786820ebae4bef04a9d67d25dd08e88)]:
+  - effect@2.4.18
+  - @effect/platform@0.48.25
+  - @effect/platform-node-shared@0.3.25
+
 ## 0.45.26
 
 ### Patch Changes