+++ packages/printer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @effect/printer
 
+## 0.32.3
+
+### Patch Changes
+
+- Updated dependencies [[`dadc690`](https://github.com/Effect-TS/effect/commit/dadc6906121c512bc32be22b52adbd1ada834594)]:
+  - effect@2.4.18
+  - @effect/typeclass@0.23.18
+
 ## 0.32.2
 
 ### Patch Changes